+++ Fluent.Ribbon/Controls/Ribbon.cs
@@ -1921,7 +1921,7 @@ namespace Fluent
                 && this.TabControl.SelectedIndex == -1
                 && this.TabControl.IsMinimized == false)
             {
-                this.TabControl.SelectedItem = this.TabControl.GetFirstVisibleItem();
+                this.TabControl.SelectedItem = this.TabControl.GetFirstVisibleAndEnabledItem();
             }
         }