+++ Changelog.md
@@ -151,6 +151,8 @@
   - [#457](../../issues/457) - Using the keytip shortcut to open a backstage briefly opens it, then closes it immediately
   - [#459](../../issues/459) - Label of the Spinner disappears, if there isn't enough place
   - [#463](../../issues/463) - ShowQuickAccessToolBarAboveRibbon="False" causes crash
+  - [#465](../../issues/465) - Clicking on ContextualGroup background or scrolling through ribbon will open disabled tab  
+    During this fix `RibbonTabControl.GetFirstVisibleItem` was renamed to `RibbonTabControl.GetFirstVisibleAndEnabledItem`.
 
 - ### Enhancements
   - `LayoutTransform` and `RenderTransform` can now be used directly on `RibbonWindow` as this now gets forwarded to the first template child of the window. Have a look at `TestWindow` in the showcase application for an example on how to use it. This was added as the fix for [#430](../../issues/430).